pyproject.toml
[project]
name = "chatbi"
version = "0.2.0"
dependencies = [
"fastapi",
"uvicorn[standard]",
"pydantic>=2",
"sqlalchemy",
"sqlmodel",
"psycopg[binary]",
"qdrant-client",
"sentence-transformers",
"langgraph",
"langchain-text-splitters",
"tenacity",
"python-dotenv",
"pytest",
]